// frontend/src/generated/endpoints/auth/auth.ts
export type loginResponse200 = {
  data: EnvelopeAuthMessage
  status: 200
}

export type loginResponseDefault = {
  data: ErrorModel
  status: Exclude<HTTPStatusCodes, 200>
}

export type loginResponseSuccess = (loginResponse200) & {
  headers: Headers;
};
export type loginResponseError = (loginResponseDefault) & {
  headers: Headers;
};

export type loginResponse = (loginResponseSuccess | loginResponseError)
// ...
    export type logoutResponse204 = {
  data: void
  status: 204
}

export type logoutResponseDefault = {
  data: ErrorModel
  status: Exclude<HTTPStatusCodes, 204>
}

export type logoutResponseSuccess = (logoutResponse204) & {
  headers: Headers;
};
export type logoutResponseError = (logoutResponseDefault) & {
  headers: Headers;
};

export type logoutResponse = (logoutResponseSuccess | logoutResponseError)